requirements.txt
sam3
google-genai
pillow
requests
aiohttp
tqdm
matplotlib
transformers
torch
python-dotenv